+++ src/pages/Public/RedirectAuth.tsx
@@ -1,18 +1,12 @@
 import { useEffect } from 'react';
 import { useNavigate, useSearchParams } from 'react-router-dom';
 
-import { useSetRecoilState } from 'recoil';
 import { useQuery } from '@tanstack/react-query';
-import { UserInfoState } from '@/stores/userInfo';
-import OAuthState from '@/stores/auth';
-
 import { getAuthMemberData, RedirectAuthTypes } from '@/api/redirectAuth';
-import { getUserInfo, silentRefresh } from '@/api/testApi';
+import useLogin from '@/hooks/useLogin';
 
 const RedirectAuth = () => {
   const navigate = useNavigate();
-  const setIsOAuth = useSetRecoilState(OAuthState);
-  const setUserInfoState = useSetRecoilState(UserInfoState);
 
   const [searchParams] = useSearchParams();
   const provider = searchParams.get('provider')!;
@@ -20,23 +14,13 @@ const RedirectAuth = () => {
 
   const { data } = useQuery<RedirectAuthTypes>(['auth'], () => getAuthMemberData(provider, code));
 
-  const saveUserInfo = async () => {
-    const { id, email, nickname, profileImage } = await getUserInfo();
-    setUserInfoState({ id, email, nickname, profileImage });
-  };
-
-  const login = async () => {
-    await silentRefresh();
-    await saveUserInfo();
-    setIsOAuth(true);
-
-    navigate('/issues');
-  };
+  const { login } = useLogin();
 
   useEffect(() => {
     const { signInMember } = data!;
     if (signInMember) {
       login();
+      navigate('/issues');
     } else {
       navigate('/signup-oauth');
     }
